doctor -> practitioner refactoring
