+++ src/main/resources/templates/practitioner/dashboard.html
@@ -7,7 +7,7 @@
   <h2>Welcome to Doctor Dashboard</h2>
   <p th:text="'Logged in as: ' + ${doctor.firstName} + ' ' + ${doctor.lastName}" />
   <p th:text="'Has role: ' + ${doctor.role}" />
-  <p><a th:href="@{/doctor/dashboard/generatePatient}">Add patient</a></p>
+  <p><a th:href="@{/practitioner/dashboard/generatePatient}">Add patient</a></p>
   <p><a th:href="@{/logout}">Log out</a></p>
 </body>
 </html>
