+++ src/main/resources/templates/practitioner/dashboard/error.html
@@ -6,8 +6,8 @@
 </head>
 <body>
 	<h1>ERROR: Patient ID does not exist, no valid condition entry or patient already existing.</h1>
-    <p><a href="/doctor/dashboard/generatePatient">Add patient</a></p>
-    <p><a th:href="@{/doctor/dashboard}">Home</a></p>
+    <p><a th:href="@{/practitioner/dashboard/generatePatient}">Add patient</a></p>
+    <p><a th:href="@{/templates/doctor/dashboard}">Home</a></p>
     <p><a th:href="@{/logout}">Log out</a></p>
 </body>
 </html>
