+++ src/main/resources/templates/practitioner/dashboard/generatePatient.html
@@ -6,11 +6,11 @@
 </head>
 <body>
 	<h1>Add new patient below by entering ID.</h1>
-    <form action="#" enctype="multipart/form-data" th:action="@{/doctor/dashboard/generatePatient}" th:object="${generatePatient}" method="post">
+    <form action="#" enctype="multipart/form-data" th:action="@{/practitioner/dashboard/generatePatient}" th:object="${generatePatient}" method="post">
     	<p>Id: <input type="text" th:field="*{id}" /></p>
         <p>File to upload (.pdf only): <input type="file" th:field="*{file}" /></p>
         <p><input type="submit" value="Add patient" /> <input type="reset" value="Reset" /></p>
     </form>
-    <p><a href="/doctor/dashboard">Home</a></p>
+    <p><a href="/practitioner/dashboard">Home</a></p>
 </body>
 </html>
