+++ src/main/resources/templates/practitioner/dashboard/patientGenerated.html
@@ -8,8 +8,8 @@
 	<h1>Patient created.</h1>
     <p th:text="'ID: ' + ${patient.id}" />
     <p th:text="'Password: ' + ${patient.password}" />
-    <p><a href="/doctor/dashboard/generatePatient">Add another patient</a></p>
-    <p><a href="/doctor/dashboard">Home</a></p>
+    <p><a href="/practitioner/dashboard/generatePatient">Add another patient</a></p>
+    <p><a href="/practitioner/dashboard">Home</a></p>
     <p><a th:href="@{/logout}">Log out</a></p>
 </body>
 </html>
